@@ -1,14 +1,5 @@
-import 'bootstrap/dist/css/bootstrap.css';
-import BootstrapVue from 'bootstrap-vue';
-import Vue from 'vue';
+import { createApp } from 'vue';
 import App from './App.vue';
 import router from './router';
 
-Vue.use(BootstrapVue);
-
-Vue.config.productionTip = false;
-
-new Vue({
-  router,
-  render: h => h(App),
-}).$mount('#app');
+createApp(App).use(router).mount('#app');
